requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pillow==10.0.1
opencv-python==4.8.1.78
imagehash==4.3.1
numpy==1.24.3
tqdm==4.66.1